lightmap atlasing
works on PC, needs to be enabled by setting LIGHT_MAP_ALIAS to 1, and
switching the commented and uncommented line in the BrushFragment
shader.

This should work on Raspberry Pi also, but I don't have access to one
for the next two weeks. Will enable by default once i have checked it.