add water warping
A vertex effect, but without the issues from the original GL quake. i.e.
No holes in the map.

This CL also contains the guts of a "render to FBO" pass so that we
could apply the water warping effect in the same way  as software quake.
It's disabled by default as the raspberry model 1 does not have the
memory for it at higher resolutions. (and/or I don't have time to try
and find the memory for it).

RPI FBO textures need to be powers of two, at 1080p I would need a
2048x2048 texture for this effect and my tests show that it would
probably need to be double buffered. So no bueno. On a model 2 raspberry
pi this would be simple, so the code is left in for future reference. I
didn't want to complicate the code further and check for models and
memory splits to enable this effect. If i were to do a Qurp+ version
targeting model 2 and higher, this would simplify some of the shaders as
I could apply the gamma once at the end. Thw water warping effect could
also be done as in the original software rendered quake.